+++ benchmark/bench_nsb_action/bencher/bencher.go
@@ -33,7 +33,7 @@ func (tasker *task) nsbRoutine(
 	signer uip.Signer, index int, txpadding []byte) {
 
 	// info, err := cli.GetAbciInfo()
-	iscAddress, err := cli.CreateISC(signer, []uint32{0}, [][]byte{signer.GetPublicKey()}, nil, txpadding)
+	iscAddress, err := cli.CreateISC(signer, []uint64{0}, [][]byte{signer.GetPublicKey()}, nil, txpadding)
 	if err != nil {
 		tasker.badSession.Add(1)
 		fmt.Println(err)
